Move CredentialIssuanceAction and util functions used by issuance FSMs
