+++ aries_vcx/src/handlers/issuance/holder/holder.rs
@@ -2,11 +2,11 @@ use std::collections::HashMap;
 
 use crate::error::prelude::*;
 use crate::handlers::connection::connection::Connection;
-use crate::handlers::issuance::actions::CredentialIssuanceAction;
 use crate::handlers::SendClosure;
 use crate::messages::a2a::A2AMessage;
 use crate::messages::issuance::credential_offer::CredentialOffer;
 use crate::messages::issuance::credential_proposal::CredentialProposalData;
+use crate::protocols::issuance::actions::CredentialIssuanceAction;
 use crate::protocols::issuance::holder::state_machine::HolderSM;
 
 #[derive(Serialize, Deserialize, Debug, Clone, Default)]
